+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - My Groups widget not displaying on home page [#1021](https://github.com/rokwire/illinois-app/issues/1021)
 - Disable local data backup on Android [#1019](https://github.com/rokwire/illinois-app/issues/1019)
 - Use different values for storage encryption key and IV [#1016](https://github.com/rokwire/illinois-app/issues/1016)
+### Added
+- Acknowledge FCM messages to redirect user to Home panel and Inbox [#1024](https://github.com/rokwire/illinois-app/issues/1024).
 
 ## [3.0.53] - 2021-11-24
 ### Changed